[REFACTOR] move plugins' config to single files
Use a single file with the same name of the plugin's config folder.
This makes it easier to find and manipulate the config files. It would
also help tidying up the overall structure.
