+++ lua/core/barbar.lua
