+++ lua/core/nvimtree.lua
