--- lua/lv-dap/config.lua
@@ -1,9 +1,0 @@
-O.plugin.dap = {
-  active = false,
-  breakpoint = {
-    text = "",
-    texthl = "LspDiagnosticsSignError",
-    linehl = "",
-    numhl = "",
-  },
-}
