--- lua/lv-floatterm/config.lua
@@ -1,10 +1,0 @@
-O.plugin.floatterm = {
-  active = false,
-  dimensions = {
-    height = 0.9,
-    width = 0.9,
-    x = 0.5,
-    y = 0.3,
-  },
-  border = "single", -- or 'double'
-}
