--- lua/lv-gitsigns/init.lua
@@ -1,11 +1,0 @@
-local M = {}
-
-M.config = function()
-  local status_ok, gitsigns = pcall(require, "gitsigns")
-  if not status_ok then
-    return
-  end
-  gitsigns.setup(O.plugin.gitsigns)
-end
-
-return M
